+++ 3-networks/modules/restricted_shared_vpc/versions.tf
@@ -16,4 +16,22 @@
 
 terraform {
   required_version = ">= 0.13"
+  required_providers {
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 3.50"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
+      version = ">= 3.50"
+    }
+  }
+
+  provider_meta "google" {
+    module_name = "blueprints/terraform/terraform-example-foundation:restricted_shared_vpc/v2.1.1"
+  }
+
+  provider_meta "google-beta" {
+    module_name = "blueprints/terraform/terraform-example-foundation:restricted_shared_vpc/v2.1.1"
+  }
 }