+++ 4-projects/modules/single_project/versions.tf
@@ -26,4 +26,12 @@ terraform {
       version = ">= 3.50"
     }
   }
+
+  provider_meta "google" {
+    module_name = "blueprints/terraform/terraform-example-foundation:single_project/v2.1.1"
+  }
+
+  provider_meta "google-beta" {
+    module_name = "blueprints/terraform/terraform-example-foundation:single_project/v2.1.1"
+  }
 }